move getTracks function to utils file
