+++ src/renderer/hooks/useOpenGovMessagePorts.ts
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { Config as ConfigOpenGov } from '@/config/processes/openGov';
+import { getTracks } from '../utils/openGov';
 import { useEffect } from 'react';
-import { getTracks } from '@/model/Track';
 import { useTracks } from '@app/contexts/openGov/Tracks';
 import { useReferenda } from '../contexts/openGov/Referenda';
 import { useTreasury } from '../contexts/openGov/Treasury';
